refactor(RHINENG-25881): EventManager to new file
